readfile is ok

+++ hal.c
@@ -18,7 +18,7 @@ size_t HAL_fread(void *ptr, size_t size, size_t count)
     return fread(ptr, size, count, g_stream);
 }
 
-int HAL_fseek(long offset)
+int HAL_fseek(uint32_t offset)
 {
     return fseek(g_stream, offset, SEEK_SET);
 }
